Save all flobs.

Rename flobs from a SQLite database and save to directory.

save_all_flobs(table_name = NULL, conn, dir = ".", sep = "_-_")

Arguments

table_name
A vector of character strings indicating names of tables to save flobs from. By default all tables are included.

conn
A SQLite connection object.

dir
A character string of the path to the directory to save files to.

sep
A string of the separator used to construct file names from values.

Value

An invisible named list of named vectors of the file names and new file names saved.

Examples

flob <- flobr::flob_obj
conn <- DBI::dbConnect(RSQLite::SQLite(), ":memory:")
DBI::dbGetQuery(conn, "CREATE TABLE Table1 (IntColumn INTEGER PRIMARY KEY NOT NULL)")

#> Warning: SQL statements must be issued with dbExecute() or dbSendStatement() instead of dbGetQuery() or dbSendQuery().

#> data frame with 0 columns and 0 rows

DBI::dbWriteTable(conn, "Table1", data.frame(IntColumn = c(1L, 2L)), append = TRUE)
key <- data.frame(IntColumn = 2L)
write_flob(flob, "BlobColumn", "Table1", key, conn, exists = FALSE)
dir <- tempdir()
save_all_flobs(conn = conn, dir = dir)

#> Table name: 'Table1'

#> Column name: 'BlobColumn'

#> Saving files to '/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpZvHt7e/Table1/BlobColumn'

#> ✖ Row 1: no file found

#> ✔ Row 2: file flobr.pdf renamed to 2.pdf

#> 

DBI::dbDisconnect(conn)